Add support for overriding uid/gid via Docker ENV
Currently overriding the UID/GID that the subsonic process will run from
the docker environment is not possible.
This creates problems when the the dir containing the music library
cannot be chowned for some reason (e.g. other processes requiring
access as well).

To remove this limitation, this commit introduces Docker environment
variables `SUBSONIC_UID` and `SUBSONIC_GID` and moves the actual
user/group creation into the entrypoint script, using the environment
variables instead of hard coded values.
This is possible, because the presence of the user is not strictly
required for the operations in the Dockerfile.

The default values for `SUBSONIC_UID` and `SUBSONIC_GID` are set to
1000, so the overall default behavior shouldn't change.

Instead, this even fixes the issue, where the subsonic user's hardcoded
homedir might diverge from the process' homedir set via the
`SUBSONIC_HOME` environment variable.