Move env var definitions to Dockerfile
Moving the env var definitions into the Dockerfile makes it IMHO clearer
that these variables are in fact part of the image's "API".

This also enables discovering the vars with `docker inspect`, without
looking into the image's entrypoint script.

Another side effect is, that we can avoid code duplication of default
values (e.g. port). This is not a big deal, but could help with future
additions.